chat/requirements.txt
fastapi
uvicorn[standard]
websockets
jinja2
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]